Refactor  in anticipation of landing GCP support
